refactor(viewer): CSS and JS cleanup
CSS:
- Replace undefined --bg-primary with --bg-void
- Remove incorrect hardcoded fallbacks from upload section (variables
  are always defined in :root)
- Extract repeated card gradient to --bg-card-gradient CSS variable
- Remove 4 redundant box-sizing: border-box declarations (already
  set on *)
- Remove deprecated -webkit-overflow-scrolling: touch
- Remove redundant -webkit-mask-* prefixes (unprefixed already present)
- Remove unnecessary !important on .upload-or font-size

JS:
- Extract duplicated expandLink/selectButton into shared chart_controls.js
  module, imported by both script.js (Group component) and
  section_views.js (cgroup renderer)
